Refactored ZKP tests
