+++ .github/workflows/docker-image.yml
@@ -13,5 +13,7 @@ jobs:
     - uses: actions/checkout@v3
       with:
         submodules: recursive
-    - name: Build the Docker image
+    - name: Build the Docker secp256k1 image
       run: docker build .
+    - name: Build the Docker zkp image
+      run: docker build -f Exhaustive/zkp/Dockerfile .
