+++ .swiftformat
@@ -42,6 +42,8 @@
 --exclude Sources/**/RNG_boring.swift
 --exclude Sources/**/SecureBytes.swift
 --exclude Sources/**/Signature.swift
+--exclude Exhaustive/**/secp256k1Tests.swift
+--exclude Exhaustive/**/XCTestManifests.swift
 
 # rules
 # --disable elseOnSameLine,semicolons
