+++ Exhaustive/zkp/Tests/zkpTests/XCTestManifests.swift
@@ -1,0 +1,1 @@
+../../../../Tests/secp256k1Tests/XCTestManifests.swift
