+++ Exhaustive/zkp/Tests/zkpTests/secp256k1Tests.swift
@@ -1,0 +1,1 @@
+../../../../Tests/secp256k1Tests/secp256k1Tests.swift
