+++ Exhaustive/zkp/Tests/zkpTests/zkpTests.swift
@@ -1,0 +1,15 @@
+import XCTest
+@testable import zkp
+
+final class zkpTests: XCTestCase {
+    func testExample() throws {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct
+        // results.
+//        XCTAssertEqual(zkp().text, "Hello, World!")
+    }
+
+    static var allTests = [
+        ("testExample", testExample)
+    ]
+}
