+++ Tests/secp256k1Tests/XCTestManifests.swift
@@ -1,11 +1,17 @@
 import XCTest
 
-@_exported import secp256k1
-
-#if !canImport(ObjectiveC)
+#if !canImport(ObjectiveC) && canImport(secp256k1)
     public func allTests() -> [XCTestCaseEntry] {
         [
             testCase(secp256k1Tests.allTests)
+        ]
+    }
+
+#elseif !canImport(ObjectiveC) && canImport(zkp)
+    public func allTests() -> [XCTestCaseEntry] {
+        [
+            testCase(secp256k1Tests.allTests),
+            testCase(zkpTests.allTests)
         ]
     }
 #endif
