--- Tests/zkpTests/XCTestManifests.swift
@@ -1,10 +1,0 @@
-import XCTest
-@_exported import zkp
-
-#if !canImport(ObjectiveC)
-    public func allTests() -> [XCTestCaseEntry] {
-        [
-            testCase(zkpTests.allTests)
-        ]
-    }
-#endif
